facial_client/real/requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bidict==0.21.2
bleach==3.2.1
boto3==1.16.59
botocore==1.19.59
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
#cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
#chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
click==7.1.2
colorama==0.4.4
colorlover==0.3.0
#conda==4.9.2
#conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
#cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cufflinks==0.17.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.7
docutils==0.17
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-SocketIO==5.0.1
gast==0.3.3
gdown==3.12.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
#idna @ file:///tmp/build/80754af9/idna_1593446292537/work
image==1.5.33
imageio==2.9.0
imagezmq==1.1.1
importlib-metadata==3.10.0
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
iso8601==0.1.14
itsdangerous==1.1.0
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.1
logging==0.4.9.6
logging-tree==1.8.1
m3u8==0.8.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
#mmdnn @ git+https://github.com/Microsoft/MMdnn.git@19562a381c27545984a216eda7591430e274e518
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.6
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.3.0.36
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkginfo==1.7.0
plotly==4.14.1
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
#pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.7.3
pymongo==3.11.2
#pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pyrealsense2==2.44.0.3073
pyrsistent==0.17.3
#PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2020.5
PyWavelets==1.1.1
pyzmq==20.0.0
readme-renderer==29.0
#requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rfc3986==1.4.0
rsa==4.6
ruamel-yaml==0.15.87
s3transfer==0.3.4
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
SecretStorage==3.3.1
Send2Trash==1.5.0
shortuuid==1.0.1
#six @ file:///tmp/build/80754af9/six_1605205327372/work
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.17
toml==0.10.2
#torch==1.8.0
#tornado==6.1
#tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
tradingview-ta==3.2.2
traitlets==5.0.5
twine==3.4.1
typing-extensions==3.7.4.3
#urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
zipp==3.4.1
